Move annotations into site preview focus toolbar
