+++ apps/ui/src/ui-desks/desk/index.tsx
@@ -1,8 +1,6 @@
 import { __ } from '@wordpress/i18n';
 import { useState } from 'react';
 import { useUpdateDeskSettings } from '@/data/queries/use-desk-config';
-import { AnnotationCommentDialog } from '@/ui-desks/annotations/comment-dialog';
-import { AnnotationsProvider, useAnnotations } from '@/ui-desks/annotations/context';
 import { Chats, ChatsProvider } from '../chats';
 import { DeskChrome } from '../chrome';
 import { DeskSettingsModal } from '../chrome/settings-modal';
@@ -86,82 +84,64 @@ function DeskShell( {
 
 	return (
 		<ChatsProvider siteId={ siteId }>
-			<AnnotationsProvider>
-				<Chats siteId={ siteId } />
-				<main
-					className={ styles.root }
-					aria-label={ getDeskLabel( siteId ) }
-					data-site-id={ siteId }
-					data-toolbar-editing={ editingToolbar ? 'true' : 'false' }
-				>
-					<DeskChrome
-						siteId={ siteId }
-						siteMapOpen={ siteMapOpen }
-						siteMapPageCount={ siteMapPageCount }
-						settingsOpen={ settingsOpen }
-						editingToolbar={ editingToolbar }
-						onToggleSiteMap={ onToggleSiteMap }
-						onToggleSettings={ () => setSettingsOpen( ( open ) => ! open ) }
-					/>
-					{ children }
-					{ siteMapIsLoading && <SiteMapLoadingWidget /> }
-					<DeskWidgetToolbar />
-					<AnnotationCommentDialogSlot />
-					<DeskSettingsModal
-						open={ settingsOpen }
-						onOpenChange={ setSettingsOpen }
-						onEditToolbar={ () => setEditingToolbar( true ) }
-					/>
-					{ editingToolbar && (
-						<>
-							<button
+			<Chats siteId={ siteId } />
+			<main
+				className={ styles.root }
+				aria-label={ getDeskLabel( siteId ) }
+				data-site-id={ siteId }
+				data-toolbar-editing={ editingToolbar ? 'true' : 'false' }
+			>
+				<DeskChrome
+					siteId={ siteId }
+					siteMapOpen={ siteMapOpen }
+					siteMapPageCount={ siteMapPageCount }
+					settingsOpen={ settingsOpen }
+					editingToolbar={ editingToolbar }
+					onToggleSiteMap={ onToggleSiteMap }
+					onToggleSettings={ () => setSettingsOpen( ( open ) => ! open ) }
+				/>
+				{ children }
+				{ siteMapIsLoading && <SiteMapLoadingWidget /> }
+				<DeskWidgetToolbar />
+				<DeskSettingsModal
+					open={ settingsOpen }
+					onOpenChange={ setSettingsOpen }
+					onEditToolbar={ () => setEditingToolbar( true ) }
+				/>
+				{ editingToolbar && (
+					<>
+						<button
+							type="button"
+							className={ styles.toolbarEditBackdrop }
+							aria-label={ __( 'Exit toolbar editing' ) }
+							onClick={ () => setEditingToolbar( false ) }
+						/>
+						<div className={ styles.toolbarEditActions }>
+							<Button
 								type="button"
-								className={ styles.toolbarEditBackdrop }
-								aria-label={ __( 'Exit toolbar editing' ) }
+								label={ __( 'Done' ) }
+								variant="chrome"
+								size="large"
 								onClick={ () => setEditingToolbar( false ) }
-							/>
-							<div className={ styles.toolbarEditActions }>
-								<Button
-									type="button"
-									label={ __( 'Done' ) }
-									variant="chrome"
-									size="large"
-									onClick={ () => setEditingToolbar( false ) }
-								>
-									{ __( 'Done' ) }
-								</Button>
-								<Button
-									type="button"
-									label={ __( 'Reset' ) }
-									variant="chrome"
-									size="large"
-									onClick={ () =>
-										updateDeskSettings( { toolbarLayout: DEFAULT_DESK_TOOLBAR_LAYOUT } )
-									}
-								>
-									{ __( 'Reset' ) }
-								</Button>
-							</div>
-						</>
-					) }
-				</main>
-			</AnnotationsProvider>
+							>
+								{ __( 'Done' ) }
+							</Button>
+							<Button
+								type="button"
+								label={ __( 'Reset' ) }
+								variant="chrome"
+								size="large"
+								onClick={ () =>
+									updateDeskSettings( { toolbarLayout: DEFAULT_DESK_TOOLBAR_LAYOUT } )
+								}
+							>
+								{ __( 'Reset' ) }
+							</Button>
+						</div>
+					</>
+				) }
+			</main>
 		</ChatsProvider>
-	);
-}
-
-function AnnotationCommentDialogSlot() {
-	const { pendingAnnotation, confirmPendingAnnotation, cancelPendingAnnotation } = useAnnotations();
-	if ( ! pendingAnnotation ) {
-		return null;
-	}
-
-	return (
-		<AnnotationCommentDialog
-			payload={ pendingAnnotation.payload }
-			onAdd={ confirmPendingAnnotation }
-			onCancel={ cancelPendingAnnotation }
-		/>
 	);
 }
